requirements.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.5.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.16
et-xmlfile==1.1.0
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
idna==3.6
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
nanoid==2.0.0
narwhals==1.9.3
numpy==1.26.2
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzbar==0.1.9
referencing==0.35.1
requests==2.31.0
rich==13.9.2
rpds-py==0.20.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.5.1
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.6
zipp==3.20.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1